Completed main translation infrastructure and translated Heading of homepage.
