+++ packages/www/app/[locale]/page.tsx
@@ -6,7 +6,7 @@ import {
   isSupportedLanguage,
 } from '../i18n/settings';
 import { createTranslation } from '../i18n/server';
-import { NextPageProps } from '../layout';
+import { NextPageProps } from './layout';
 
 export default async function Home({ params: { locale } }: NextPageProps) {
   const lang = isSupportedLanguage(locale)
@@ -16,11 +16,12 @@ export default async function Home({ params: { locale } }: NextPageProps) {
   return (
     <div className="relative flex flex-col items-center justify-center gap-4 py-8 md:py-10">
       <Heading locale={lang} />
-      <div className="my-6">{t('heading.title')}</div>
 
       {/* web3 allowing people to share in ownershipt and governance */}
 
-      {/* <h3>What is this about?</h3>
+      <h3>{t('about.title')}</h3>
+
+      {/* 
       <p>
         Quilombo is not just a platform; it&apos;s a movement towards digitally
         empowering the Capoeira community.
