+++ packages/www/app/i18n/locales/de/home.json
@@ -1,7 +1,10 @@
 {
   "heading": {
     "title": "Willkommen bei Axé DAO",
-    "subtitle": "A digital organization for the global Capoeira Community",
-    "introduction": "The advent of <i>web3</i> has given the world a new set of tools. We are Capoeiristas from different corners of the world bringing these tools to the Capoeira community to create shared value. It is time for Capoeira to build its own open infrastructure and participate in the digital asset economy while giving every member a voice in shaping the future of this art form we love.<br /><br />Like Capoeira, the cypherpunk culture from which <i>crypto/web3</i> originates is rooted in the spirit of autonomy, resistance, freedom and self-determination. Axé DAO is a pioneering effort by Capoeiristas for Capoeiristas to channel collective resources into efforts that benefit Capoeira guided by the principles of decentralization and collective action."
+    "subtitle": "Eine digitale Organisation für die globale Capoeira Community",
+    "introduction": "Wir sind Capoeiristas aus unterschiedlichen Ecken der Welt, die die Innovationen und Werkzeuge des <i>web3</> der Capoeira Community näher bringen, um Gemeinwohl zu erzeugen. Es ist Zeit für Capoeira, ihre eigene, offene Infrastruktur aufzubauen inklusive demokratischer Entscheidungsprozeese, um die Zukunft dieser Kunstform mitzugestalen, die wir so lieben.<br /><br />Ähnlich der Capoeira ist die Cypherpunk-Kultur, aus der <i>web3</i> entsprang, verwurzelt im Geiste von Autonomie, Widerstand, Freiheit und Selbstbestimmung. Axé DAO ist ein wegweisendes Projekt von Capoeiristas für Capoeiristas, um gemeinsame Resourcen und Anstrengungen zu bündeln, die zum Wohle der Capoeira beitragen und von Prinzipien der Dezentralisation und kollektiven Handelns geleitet sind."
+  },
+  "about": {
+    "title": "Worum geht's?"
   }
 }
