+++ packages/www/app/i18n/locales/pt/home.json
@@ -1,7 +1,10 @@
 {
   "heading": {
     "title": "Bemvindo a Axé DAO",
-    "subtitle": "A digital organization for the global Capoeira Community",
-    "introduction": "The advent of <i>web3</i> has given the world a new set of tools. We are Capoeiristas from different corners of the world bringing these tools to the Capoeira community to create shared value. It is time for Capoeira to build its own open infrastructure and participate in the digital asset economy while giving every member a voice in shaping the future of this art form we love.<br /><br />Like Capoeira, the cypherpunk culture from which <i>crypto/web3</i> originates is rooted in the spirit of autonomy, resistance, freedom and self-determination. Axé DAO is a pioneering effort by Capoeiristas for Capoeiristas to channel collective resources into efforts that benefit Capoeira guided by the principles of decentralization and collective action."
+    "subtitle": "Uma organização digital para a comunidade global de Capoeira",
+    "introduction": "Somos capoeiristas de diferentes partes do mundo trazendo as inovações e ferramentas da <i>web3</i> para a comunidade capoeirista para criar um valor compartilhado. É hora da Capoeira construir sua própria infraestrutura aberta com propriedade e governança compartilhadas para moldar o futuro dessa arte que amamos.<br /><br />Assim como a Capoeira, a cultura cypherpunk da qual a <i>web3</i> se origina é fundamentada no espírito de autonomia, resistência, liberdade e autodeterminação. Axé DAO é um esforço pioneiro de Capoeiristas para Capoeiristas para canalizar recursos coletivos em esforços que beneficiem a Capoeira, guiados pelos princípios da descentralização e ação coletiva."
+  },
+  "about": {
+    "title": "O que é isso?"
   }
 }
